ara/requirements.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
distribute==0.6.31
django-grappelli==2.4.7
django-extensions==1.2.3
django-relatives==0.3.1
gunicorn==18.0
psycopg2==2.5.1
wsgiref==0.1.2